perf(imaging): the slow traversal was two sorts for two medians, and my "14 ns/px" was a Debug number
Asked to look at Statistics. The first thing measurement said was that the previous
measurement was wrong.

Filed yesterday in docs/todo/imaging.md: "Image.Histogram is the cost of a document open,
and it is ~14 ns/px", from Statistics(c) x3 = 1,028-1,195 ms on a 24 MP 3-channel probe.
That was a DEBUG run -- dotnet test defaults to Debug and these inner loops are ~7x slower
there, so the figure described the test configuration, not the product. In Release the same
probe reports 2.08 ns/px and the ranking inverts:

  stage (24 MP, 3ch)                          Debug (as filed)  Release before  after
  Statistics(c) x3                             1,028-1,195 ms    130-177 ms      108-152 ms
  GetStarMaskedMedianAndMADScaledToUnit(c) x3    557-755 ms      412-632 ms       23-71 ms

So the hot spot was the star-masked median, and its cause was algorithmic rather than micro:
TWO full Array.Sort calls over ~1.5 M samples per channel to extract two medians. A median
needs selection, not a sort.

Fixed by reusing what was already here. StatisticsHelper had a private QuickSelect behind
MedianFast/PercentileFast; this adds a public NthSmallest over it, and the star-masked path
does two selections over ONE buffer (dropping the second 6 MB allocation per channel too).
Normalizer.MedianViaQuickSelect had already made exactly this trade for the stacking hot path
("~150 ms per channel, was ~1.5-2 s with sort-based path"), and Image.Histogram.cs already
had `using static StatisticsHelper` while still sorting.

NthSmallest rather than MedianFast is load-bearing: sorted[count / 2] is the UPPER median for
even count, MedianFast averages the two middle values, and this result feeds the stretch.

The histogram loop went 50 -> 32 ms per 24 MP channel from two changes a code reading ranks
the wrong way round -- a flat span instead of float[,] [h, w] indexing (50 -> 36), and a
float-domain clamp because Math.Clamp(float, int, uint) binds the DOUBLE overload, so the old
form ran float -> double -> clamp -> double -> int per pixel (36 -> 32). A further 32 -> 8 ms
is available by parallelising row bands and is deliberately NOT taken: it reorders the
total_value summation that feeds Background()'s mode search and hence the detection threshold,
so it cannot carry the bit-identical claim the other three do. Filed with its number.

ImageHistogram.Total was a pixel COUNT stored in a float, which quantises above 2^24 -- a
24 MP frame already exceeds that. Now long; long -> float is an implicit conversion, so every
existing consumer keeps today's exact arithmetic. That exposed the same shape twice more in
Image.Stretch, where cumulative bin counts were accumulated in float and so saturated at
16,777,216 mid-walk; benign on a background-dominated astro frame where halfTotal is reached
in the first few bins, wrong where the median bin is late. Also drops a dead `float total`
parameter from ComputeLogSlopeRSquared.

Verification: HistogramSelectionParityTests reimplements the PRE-change algorithm as an oracle
and demands exact equality on three real fixtures -- bins, mean, total, median and MAD.
Sabotage-validated, and one of those sabotages is worth recording: swapping the production
call to MedianFast left every fixture assertion GREEN, because a quantised background ties the
two middle samples, so no real image can distinguish the conventions. Shifting the order
statistic to count / 3 does fail. The upper-median convention is therefore pinned on synthetic
distinct values in NthSmallestTests instead, which is the only place it is observable.

Full suite green: 4,611 passed, 0 failed.

+++ src/TianWen.Lib/Imaging/ImageHistogram.cs
@@ -9,7 +9,10 @@ namespace TianWen.Lib.Imaging;
 /// <param name="Channel">Which channel this histogram was generated for</param>
 /// <param name="Histogram">histogram values</param>
 /// <param name="Mean"></param>
-/// <param name="Total"></param>
+/// <param name="Total">Number of pixels that landed in the histogram. A COUNT, so an integer
+/// type: as a <c>float</c> it was quantised above 2^24, which a 24 MP frame already exceeds.
+/// <c>long</c> converts implicitly to <c>float</c>/<c>double</c>, so consumers doing fractional
+/// arithmetic on it are unaffected.</param>
 /// <param name="Threshold"></param>
 /// <param name="ThresholdPct">Percentage of pixels above the threshold</param>
 /// <param name="RescaledMaxValue">when not null, specifies the max pixel value the image was rescaled to</param>
@@ -21,7 +24,7 @@ public record ImageHistogram(
     int Channel,
     ImmutableArray<uint> Histogram,
     float Mean,
-    float Total,
+    long Total,
     float Threshold,
     byte ThresholdPct,
     float? RescaledMaxValue,
